+++ app/app.go
@@ -131,6 +131,8 @@ import (
 	spotkeeper "github.com/NibiruChain/nibiru/x/spot/keeper"
 	spottypes "github.com/NibiruChain/nibiru/x/spot/types"
 
+	"github.com/NibiruChain/nibiru/x/sudo"
+
 	"github.com/NibiruChain/nibiru/x/stablecoin"
 	stablecoinkeeper "github.com/NibiruChain/nibiru/x/stablecoin/keeper"
 	stablecointypes "github.com/NibiruChain/nibiru/x/stablecoin/types"
@@ -196,6 +198,7 @@ var (
 		perp.AppModuleBasic{},
 		perpamm.AppModuleBasic{},
 		inflation.AppModuleBasic{},
+		sudo.AppModuleBasic{},
 		wasm.AppModuleBasic{},
 		ibcfee.AppModuleBasic{},
 	)
@@ -219,6 +222,7 @@ var (
 		perptypes.FeePoolModuleAccount:        {},
 		epochstypes.ModuleName:                {},
 		stablecointypes.StableEFModuleAccount: {authtypes.Burner},
+		sudo.ModuleName:                       {},
 		common.TreasuryPoolModuleAccount:      {},
 		wasm.ModuleName:                       {},
 	}
@@ -300,6 +304,7 @@ type NibiruApp struct {
 	OracleKeeper     oraclekeeper.Keeper
 	StablecoinKeeper stablecoinkeeper.Keeper
 	InflationKeeper  inflationkeeper.Keeper
+	SudoKeeper       sudo.Keeper
 
 	// WASM keepers
 	WasmKeeper       wasm.Keeper
@@ -335,6 +340,7 @@ func GetWasmOpts(nibiru NibiruApp, appOpts servertypes.AppOptions) []wasm.Option
 	wasmOpts = append(wasmOpts, wasmbinding.RegisterWasmOptions(
 		&nibiru.PerpKeeper,
 		&nibiru.PerpAmmKeeper,
+		&nibiru.SudoKeeper,
 	)...)
 
 	return wasmOpts